Made NodeViewModel generic and moved existing ones into PixiEditor.ViewModels.Document.Nodes
