+++ src/PixiEditor/ViewModels/Document/Nodes/FolderViewModel.cs
@@ -1,10 +1,11 @@
 ﻿using System.Collections.ObjectModel;
+using PixiEditor.ChangeableDocument.Changeables.Graph.Nodes;
 using PixiEditor.Models.DocumentModels;
 using PixiEditor.Models.Handlers;
 
-namespace PixiEditor.ViewModels.Document;
+namespace PixiEditor.ViewModels.Document.Nodes;
 #nullable enable
-internal class FolderViewModel : StructureMemberViewModel, IFolderHandler
+internal class FolderViewModel : StructureMemberViewModel<FolderNode>, IFolderHandler
 {
     public FolderViewModel()
     {
